미네랄 클래스 추가

+++ TermProject/app/src/main/java/kr/ac/kpu/sgp02/termproject/game/monster/Armor.java
@@ -27,6 +27,6 @@ public class Armor extends Monster {
 
         speed = Metrics.size(R.dimen.armor_speed);
         maxHp = Metrics.floatValue(R.dimen.armor_hp);
-        reward = Metrics.floatValue(R.dimen.armor_reward);
+        reward = Metrics.intValue(R.dimen.armor_reward);
     }
 }
